strana_bot/telegramConnect/pyproject.toml
[tool.poetry]
name = "strana-bot"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.4"
openai = "^1.35.1"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.5"
loguru = "^0.7.2"
requests = "^2.32.3"
aiogram = "^3.8.0"
ipywidgets = "^8.1.3"
psycopg2-binary = "^2.9.9"
update = "^0.0.1"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
